refactored file structure, minor html and css changes as per feedback from sprint 1
