+++ src/components/Hero/Hero.js
@@ -1,8 +1,8 @@
-import videos from "../Data/videos.json";
+import videos from "../../Data/videos.json";
 import "./Hero.scss";
-import eyeball from "../assets/Icons/views.svg";
-import heart from "../assets/Icons/likes.svg";
-import videodetails from "../Data/video-details.json";
+import eyeball from "../../assets/Icons/views.svg";
+import heart from "../../assets/Icons/likes.svg";
+import videodetails from "../../Data/video-details.json";
 
 export default function Hero({ selectedVideoDetails }) {
   const date = new Date(selectedVideoDetails.timestamp);
@@ -12,7 +12,7 @@ export default function Hero({ selectedVideoDetails }) {
   const parsedDate = month + "/" + day + "/" + year;
 
   return (
-    <div className="hero">
+    <hero className="hero">
       <video
         className="hero__video"
         controls
@@ -34,14 +34,14 @@ export default function Hero({ selectedVideoDetails }) {
           <div className="main__info--two">
           <p className="main__info--views">
               {" "}
-              <img className="main__info--views-image" src={eyeball} />
+              <img className="main__info--views-image" src={eyeball} alt=" grey oval shaped eyeball" />
               {selectedVideoDetails.views}
             </p>
           
 
             <p className="main__info--likes">
               {" "}
-              <img className="main__info--likes-image" src={heart} />
+              <img className="main__info--likes-image" src={heart} alt="grey heart shape" />
               {selectedVideoDetails.likes}
             </p>
           </div>
@@ -50,6 +50,6 @@ export default function Hero({ selectedVideoDetails }) {
           {selectedVideoDetails.description}
         </div>
       </div>
-    </div>
+    </hero>
   );
 }
