+++ src/components/Hero/Hero.scss
@@ -1,5 +1,5 @@
-@import "../styles/partials/variables";
-@import "../styles/partials/mixins";
+@import "../../styles/partials/variables";
+@import "../../styles/partials/mixins";
 .hero {
     background-color: $White;
 }
