+++ src/components/Nextvids/Nextvids.js
@@ -8,7 +8,7 @@ function Nextvids({ videos, selectedVideo, setSelectedVideo }) {
   };
 
   return (
-    <div className="video-container">
+    <section className="video-container">
       <h2 className="video-container__header">NEXT VIDEOS</h2>
       <div className="video-container__list">
         {filterVideos.map((video) => (
@@ -17,7 +17,7 @@ function Nextvids({ videos, selectedVideo, setSelectedVideo }) {
             key={video.id}
             onClick={() => videoSelector(video)}
           >
-            <img className="video-container__image" src={video.image} />
+            <img className="video-container__image" src={video.image} alt="video thumbnail" />
             <div className="video-container__details">
               <h2 className="video-container__title">{video.title}</h2>
               <p className="video-container__channel">{video.channel}</p>
@@ -25,7 +25,7 @@ function Nextvids({ videos, selectedVideo, setSelectedVideo }) {
           </div>
         ))}
       </div>
-    </div>
+    </section>
   );
 }
 export default Nextvids;
